Updated input/output

+++ src/Trading.Application/Trading.Application.BLL/CapitalIntegrationEntities/CreatePositionEntity.cs
@@ -3,5 +3,5 @@ namespace Trading.Application.BLL.CapitalIntegrationEntities;
 public class CreatePositionEntity : UpdatePositionEntity {
     public string Direction { get;set; }
     public required string Epic { get;set; }
-    public double Size { get; set; }
+    public decimal Size { get; set; }
 }
