+++ src/Trading.Application/Trading.Application/UserContext/OrderData.cs
@@ -5,8 +5,8 @@ namespace Trading.Application.UserContext;
 public class TradeData {
     public string Epic { get;set; }
     public Directions Direction { get;set; }
-    public double? Level { get;set; }
-    public double Size { get;set; }
-    public double StopLoss { get;set; }
-    public double TakeProfit { get;set; }
+    public decimal? Level { get;set; }
+    public decimal Size { get;set; }
+    public decimal StopLoss { get;set; }
+    public decimal TakeProfit { get;set; }
 }
