+++ src/Trading.Application/Trading.Application/UserInputPipeline/PipelineSteps/ParseTradeCreationStep.cs
@@ -19,10 +19,10 @@ public class ParseTradeCreationStep(IUserContext userContext, ILogger<ParseTrade
             {
                 Epic = lines[0],
                 Direction = (Directions)Enum.Parse(typeof(Directions), lines[1]),
-                Size = double.Parse(lines[2]),
-                Level = isOrder ? double.Parse(lines[3]) : null,
-                StopLoss = hasStoploss ? double.Parse(lines[isOrder ? 4 : 3]) : 0,
-                TakeProfit = hasTakeprofit ? double.Parse(lines[isOrder ? 5 : 4]) : 0,
+                Size = decimal.Parse(lines[2]),
+                Level = isOrder ? decimal.Parse(lines[3]) : null,
+                StopLoss = hasStoploss ? decimal.Parse(lines[isOrder ? 4 : 3]) : 0,
+                TakeProfit = hasTakeprofit ? decimal.Parse(lines[isOrder ? 5 : 4]) : 0,
             };
         } catch {
             logger.LogError($"Can't parse input: {input}");
