+++ src/Trading.Application/Trading.Application/UserInputPipeline/PipelineSteps/ParseTradeUpdateStep.cs
@@ -16,9 +16,9 @@ public class ParseTradeUpdateStep(IUserContext userContext, ILogger<ParseTradeUp
         try {
             userContext.OrderData = new TradeData
             {
-                Level = isOrder ? double.Parse(lines[0]) : null,
-                StopLoss = hasStoploss ? double.Parse(lines[isOrder ? 1 : 0]) : 0,
-                TakeProfit = hasTakeprofit ? double.Parse(lines[isOrder ? 2 : 1]) : 0
+                Level = isOrder ? decimal.Parse(lines[0]) : null,
+                StopLoss = hasStoploss ? decimal.Parse(lines[isOrder ? 1 : 0]) : 0,
+                TakeProfit = hasTakeprofit ? decimal.Parse(lines[isOrder ? 2 : 1]) : 0
             };
         }
         catch
